block, bfq-mq: measure __bfq_insert_request execution time
Commit 9d3206e8e116 (in branch "bfq-mq-time-deltas-v0") allows to measure
execution times in a really rough way, where the main bottlenecks are.

This commit shows how measures can be more accurate.

The main difference is the absence of many trace_printk() calls, which
would introduce a not so negligible overhead (particularly noticeable on
fast io devices).  In fact all the time deltas are now computed and
printed outside the function which contains all the time measurements.

The second difference is that the previous total time measure is now the
time delta of the root function (e.g. "__bfq_insert_request"). While the
other time deltas are the measures of its sub-function/pieces of code.
For this reasons the printed records are now shown in a tree like
representation.

For example:
. __bfq_insert_request    t_delta:  61191 ns # <- root function
`-- bfq_setup_cooperator  t_delta:    808 ns # <- sub-function
    ...     ...             ...       ...
`-- bfq_rq_enqueued       t_delta:   4979 ns # <- sub-function

As in the bfq-mq-time-deltas-v0 branch, the following command should
give you an aligned view of the measures:
sudo cat /sys/kernel/debug/tracing/trace \
| sed 's/^.*:\s*block/block/'            \
| grep '^block'

Signed-off-by: Federico Motta <federico@willer.it>